requirements.txt
alembic==1.8.1
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
dnspython==2.2.1
dominate==2.7.0
email-validator==1.3.0
Flask==2.2.2
Flask-Login==0.6.2
Flask-Migrate==3.1.0
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.0.1
greenlet==1.1.3.post0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.3
MarkupSafe==2.1.1
pathlib==1.0.1
psycopg2-binary==2.9.5
python-dotenv==0.21.0
requests==2.28.1
SQLAlchemy==1.4.42
urllib3==1.26.12
visitor==0.1.3
Werkzeug==2.2.2
WTForms==3.0.1
mercadopago==2.2.0
